factor write command tests to own module
