+++ helix-term/tests/integration.rs
@@ -23,5 +23,4 @@ mod test {
     mod movement;
     mod prompt;
     mod splits;
-    mod write;
 }
